feat(api-types): scaffold workspace crate + move shared enums
New crates/swissknife-api-types workspace member as the single source of
truth for contract types. Move Currency, Ledger, BtcNetwork, Permission
into it; old module paths now re-export from the crate. The
lightning_invoice -> Currency conversion becomes a local free function
(orphan rule).
