+++ crates/swissknife-api-types/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "swissknife-api-types"
+version = "0.1.7"
+edition = "2021"
+authors = ["Dario Anongba Varela <dario.varela@numeraire.tech>"]
+rust-version = "1.96"
+license = "MIT"
+description = "Shared API/contract types (entities, enums, requests) for Numeraire SwissKnife."
+repository = "https://github.com/bitcoin-numeraire/swissknife"
+
+[dependencies]
+serde = { version = "1.0.228", features = ["derive"] }
+strum = "0.28"
+strum_macros = "0.28"
+utoipa = { version = "5.5.0", features = ["chrono", "uuid"] }
+chrono = { version = "0.4.44", features = ["serde"] }
+uuid = { version = "1.23.2", features = ["v4", "serde"] }
